Fixes git-tfs-remote always shows usage text
This commit fixes #2 whereby running the command with valid command line
options resulted in the help and usage text still being displayed. It
should be noted that the command was still successufl, even thought it
may not have appeared that anything happened. This is because the
command follows the output pattern of the native git-remote command,
whereby if -v is not supplied, no output is generated.