Rename `git-tfs-merge` to `git-tfs-merge-into`
In git, you checkout the branch which is to receive changes from a
different branch and then issues `git merge <source-branch>`. However,
with git-tfs, this can cause problems under certain circumstances,
especially when working with a small team.

Instead, using the scripts in this repository, it's a best practice to
perform all development work in a branch off of the main/trunk branch.
This script will then "merge" this "feature branch" _into_ the specified
branch. So this is a "to" operation instead of a "from" operation (such
as in git proper).

In reality, what's really happening is that all of the changes in your
feature branch are `git-tfs rcheckin`-ed to the TFS
repository--essentially, "rebasing" your changes onto the specified TFS
branch by "replaying" them one-by-one, as opposed to actually performing
a merge (which is also possible, but my teams prefer to have a linear
history instead of a squashed merge-commit).

There are also some smarts built-in to the script to help with keeping
PRs on the shared git server "in-sync" so that they appear Merged
instead of Declined (I'm looking at you Atlassian Stash/BitBucket).

Closes #1.