Remove branch name from fetch command
This commit removes the branch name from the fetch command when synching
up with the shared git repository. The reason is that for working with
TFS, we export metadata (which gets added as commit notes) from TFS for
workitem and changeset information. When working in a team environment,
specifying the branch name to fetch will cause the refs/notes/commits
branch to not be brought down and possible lead to different team
members having different HEAD refs which will cause push conflicts.